feat(node): adopt Packemon for project management
