+++ package.json
@@ -14,23 +14,31 @@
   "license": "MIT",
   "author": "Noah Sherwin",
   "files": [
+    "!tests/",
     "CHANGELOG.md",
-    "node/",
+    "cjs/**/*",
     "dotnet/",
-    "!__tests__/",
-    "!.jest/",
-    "!.husky/",
-    "!./.github/"
+    "src/**/*",
+    "static/**/*"
   ],
-  "main": "index.js",
+  "main": "./cjs/index.cjs",
+  "infra": "polyrepo",
   "repository": "github:halospv3/hce.shared",
   "scripts": {
-    "build": "npx tsc",
-    "prestart": "npx tsc",
+    "build": "packemon build",
+    "check": "npm run type && npm run test && npm run lint",
+    "clean": "packemon clean",
+    "coverage": "npm run test --coverage",
+    "format": "prettier --write ./src ./tests",
+    "lint": "eslint --cache --fix ./src ./tests",
+    "pack": "packemon pack --addEngines --addFiles --declaration",
     "prepare": "husky install",
+    "presemantic-release": "npm run pack && npm run check",
     "semantic-release": "npx semantic-release",
-    "start": "node node/dist/index.mjs",
-    "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js"
+    "test": "node --loader tsx --no-warnings --test ./tests/index.test.mts",
+    "type": "tsc",
+    "validate": "packemon validate",
+    "watch": "packemon watch"
   },
   "dependencies": {
     "@commitlint/cli": "^17.7.1",
@@ -46,40 +54,54 @@
     "conventional-changelog-conventionalcommits": "^7.0.2",
     "husky": "^8.0.3",
     "node-fetch": "^3.3.2",
-    "js-yaml": "^4.1.0",
     "semantic-release": "^22.0.0"
   },
   "devDependencies": {
-    "@jest/globals": "^29.6.4",
     "@tsconfig/node-lts": "^18.12.4",
     "@types/debug": "^4.1.8",
-    "@types/jest": "^29.5.4",
     "@types/js-yaml": "^4.0.5",
     "@types/node": "^20.5.1",
     "@types/semantic-release": "^20.0.1",
-    "jest": "^29.6.2",
-    "ts-jest": "^29.1.1",
-    "ts-node": "^10.9.1",
+    "@typescript-eslint/eslint-plugin": "^6.7.5",
+    "@typescript-eslint/parser": "^6.7.5",
+    "babel-cli": "^6.26.0",
+    "babel-plugin-cjs-esm-interop": "^3.0.2",
+    "eslint": "^8.51.0",
+    "eslint-config-moon": "^2.0.10",
+    "node-fetch": "^3.3.2",
+    "packemon": "^3.2.1",
+    "prettier": "^3.0.3",
+    "prettier-config-moon": "^1.1.2",
+    "tsconfig-moon": "^1.3.0",
     "tslib": "^2.6.2",
-    "typescript": "^5.1.6"
+    "tsx": "^3.13.0",
+    "typescript": "^5.2.2"
   },
   "engines": {
-    "node": ">=18.17.1"
+    "node": ">=16.12.0"
   },
   "publishConfig": {
+    "exports": {
+      "node": {
+        ".": {
+          "import": "./index-wrapper.mjs",
+          "types": "./index.d.cts",
+          "require": "./index.cjs"
+        }
+      }
+    },
     "registry": "https://npm.pkg.github.com",
     "tag": "latest"
   },
+  "types": "./cjs/index.d.ts",
   "commitlint": {
     "extends": [
       "@commitlint/config-conventional"
     ]
   },
-  "// jest": "If the configurations for Jest, TS-Jest, or VSCode-Jest are changed, tests may fail if an active test runner fails to update its configuration and then defaults to the incompatible default configuration. If this occurs, try restarting the test runner before troubleshooting anything else. In VSCode, open the Testing pane, right-click the test suite, and select 'Update Snapshot'. I don't know why that works.",
-  "jest": {
-    "preset": "ts-jest/presets/js-with-ts-esm",
-    "resolver": "<rootDir>/.jest/mjs-resolver.ts",
-    "testEnvironment": "node"
+  "packemon": {
+    "platform": "node",
+    "format": "cjs"
   },
   "// semantic-release": "After HCE.Shared's Shareable Configuration (see 'node/static/releaserc.yml', https://semantic-release.gitbook.io/semantic-release/usage/configuration#extends) is loaded, the following semantic-release configuration is merged into the configuration object, overriding existing values. Later, when a release is triggered, semantic-release will add any unspecified settings with default values.",
   "// semantic-release - How to review the 'actual' config": "`npx semantic-release --dry-run --debug`. This will run semantic-release with 'dry-run' (no committing changes, tags, publishing activity, and more) and 'debug' (output verbose info, including the actual config object).",
@@ -87,14 +109,13 @@
   "// semantic-release - plugins configuration": "Plugins are configured via tuple-like arrays e.g. ['plugin name', {config object}].",
   "release": {
     "extends": [
-      "node/dist/index.mjs"
+      "./cjs/index.cjs"
     ],
     "plugins": [
       [
         "@semantic-release/commit-analyzer",
         {
           "// releaseRules - About": "https://github.com/semantic-release/commit-analyzer#releaserules",
-          "// releaseRules - ": "",
           "// defaults": "https://github.com/semantic-release/commit-analyzer/blob/master/lib/default-release-rules.js",
           "releaseRules": [
             {
